Make correct wallet and some small changes for portfolio and currency

+++ GorimoozeWallet/Controllers/WalletController.cs
@@ -1,6 +1,8 @@
 ﻿using GorimoozeWallet.Dto;
 using GorimoozeWallet.Interfaces;
+using GorimoozeWallet.Services;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
 
 namespace GorimoozeWallet.Controllers
 {
@@ -15,57 +17,85 @@ namespace GorimoozeWallet.Controllers
             _walletService = walletService;
         }
 
-        [HttpGet]
+        [HttpGet("getAll")]
         public IActionResult GetAll()
         {
-            var wallets = _walletService.GetWalletList();
+            var walletList = _walletService.GetList();
 
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            return Ok(wallets);
+            return Ok(walletList);
         }
 
-        [HttpGet("{userId}")]
-        public IActionResult GetByUserId(long userId)
+        [HttpGet("getById/{walletId}")]
+        public IActionResult GetById(long walletId)
         {
-            var wallet = _walletService.GetWalletByUserId(userId);
-
-            if (!ModelState.IsValid)
+            if (walletId == 0)
                 return BadRequest(ModelState);
 
-            return Ok(wallet);
+            if (_walletService.Exists(walletId))
+            {
+                ModelState.AddModelError("", "Portfolio was not found");
+                return BadRequest(ModelState);
+            }
+
+            var walletDto = _walletService.GetById(walletId);
+
+            return Ok(walletDto);
         }
 
-        [HttpPost]
-        public IActionResult Create([FromBody] WalletDto wallet)
+        [HttpPost("create")]
+        public IActionResult Create([FromBody] WalletDto walletDto)
+        {
+            if (walletDto == null)
+                return BadRequest(ModelState);
+
+            if (walletDto.PortfolioId != null && walletDto.CurrencyId != null)
+            {
+                _walletService.Create(walletDto);
+            }
+            else
+            {
+                ModelState.AddModelError("", "You must have Portfolio and select one Currency");
+                return BadRequest(ModelState);
+            }
+
+            return Ok("Successfully created");
+        }
+
+        [HttpPut("update/{walletId}")]
+        public IActionResult Update(long walletId, [FromBody] WalletDto walletDto)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            _walletService.Create(wallet);
-
-            return Ok();
-        }
-
-        [HttpPut("{walletId}")]
-        public IActionResult Update(long walletId, [FromBody] WalletDto wallet)
-        {
-            if (!ModelState.IsValid)
+            if (walletId != walletDto.Id)
+            {
+                ModelState.AddModelError("", "Ids are not equal");
                 return BadRequest(ModelState);
+            }
 
-            _walletService.Update(wallet);
-            return Ok();
+            if (!_walletService.Exists(walletId))
+                return NotFound();
+
+            _walletService.Update(walletId, walletDto);
+
+            return Ok("Successfully updated");
         }
 
-        [HttpDelete("{walletId}")]
+        [HttpDelete("delete/{walletId}")]
         public IActionResult Delete(long walletId)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
+            if (!_walletService.Exists(walletId))
+                return NotFound();
+
             _walletService.Delete(walletId);
-            return Ok();
+
+            return Ok("Successfully deleted");
         }
     }
 }
